IGNITE-5928 .NET: Get rid of BinaryStreamBase
